requirements.txt
requests
youtube-transcript-api
google-generativeai
